Move shared functionality to utils.go
